Rename trip.py to trip_old.py to make room for the new wrapper classes
Also deal with the resulting fallout
