+++ emission/core/wrapper/trip_old.py
