+++ emission/tests/analysisTests/TestSimilarity.py
@@ -3,7 +3,7 @@ import emission.core.get_database as edb
 import emission.analysis.modelling.tour_model.similarity as similarity
 import emission.simulation.trip_gen as tg
 import math
-from emission.core.wrapper.trip import Trip, Coordinate
+from emission.core.wrapper.trip_old import Trip, Coordinate
 import emission.analysis.modelling.tour_model.cluster_pipeline as cp
 import datetime 
 import os, os.path
